refactor: estrai domain/source_utils + domain/preflight da cli/
Rimuove ciclo di dipendenza preflight_ops ↔ cmd_run:
- _resolve_source → domain/source_utils (entrambi i consumer lo importano da lì)
- run_preflight → domain/preflight.py (con dipendenze solo core/ + domain/)
- cli/preflight_ops.py diventa re-export puro

Azzera ultima dipendenza MCP→cli/ per preflight_ops.
